feat(policies): Adds a policy detail page
